refactor(global): datagsm-common ResDto/Enum을 datagsm-shared로 마이그레이션
