--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/club/dto/internal/ClubSummaryDto.kt
@@ -1,13 +1,0 @@
-package team.themoment.datagsm.common.domain.club.dto.internal
-
-import io.swagger.v3.oas.annotations.media.Schema
-import team.themoment.datagsm.common.domain.club.entity.constant.ClubType
-
-data class ClubSummaryDto(
-    @field:Schema(description = "동아리 ID", example = "1")
-    val id: Long,
-    @field:Schema(description = "동아리 이름", example = "SW개발동아리")
-    val name: String,
-    @field:Schema(description = "동아리 종류", example = "MAJOR_CLUB")
-    val type: ClubType,
-)
