+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/club/dto/request/QueryClubReqDto.kt
@@ -4,10 +4,10 @@ import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.Max
 import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.Positive
-import team.themoment.datagsm.common.domain.club.entity.constant.ClubSortBy
-import team.themoment.datagsm.common.domain.club.entity.constant.ClubStatus
-import team.themoment.datagsm.common.domain.club.entity.constant.ClubType
 import team.themoment.datagsm.common.global.constant.SortDirection
+import team.themoment.datagsm.shared.domain.club.constant.ClubSortBy
+import team.themoment.datagsm.shared.domain.club.constant.ClubStatus
+import team.themoment.datagsm.shared.domain.club.constant.ClubType
 
 data class QueryClubReqDto(
     @field:Positive
