--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/club/entity/constant/ClubSortBy.kt
@@ -1,9 +1,0 @@
-package team.themoment.datagsm.common.domain.club.entity.constant
-
-enum class ClubSortBy {
-    ID,
-    NAME,
-    TYPE,
-    FOUNDED_YEAR,
-    STATUS,
-}
