--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/club/entity/constant/ClubStatus.kt
@@ -1,8 +1,0 @@
-package team.themoment.datagsm.common.domain.club.entity.constant
-
-enum class ClubStatus(
-    val value: String,
-) {
-    ACTIVE("운영 중"),
-    ABOLISHED("폐지"),
-}
