--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/club/entity/constant/ClubType.kt
@@ -1,8 +1,0 @@
-package team.themoment.datagsm.common.domain.club.entity.constant
-
-enum class ClubType(
-    val value: String,
-) {
-    MAJOR_CLUB("전공동아리"),
-    AUTONOMOUS_CLUB("창체동아리"),
-}
