--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/neis/meal/entity/constant/MealType.kt
@@ -1,7 +1,0 @@
-package team.themoment.datagsm.common.domain.neis.meal.entity.constant
-
-enum class MealType {
-    BREAKFAST,
-    LUNCH,
-    DINNER,
-}
