+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/project/dto/request/ProjectReqDto.kt
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Positive
 import jakarta.validation.constraints.Size
-import team.themoment.datagsm.common.domain.project.entity.constant.ProjectStatus
+import team.themoment.datagsm.shared.domain.project.constant.ProjectStatus
 
 data class ProjectReqDto(
     @field:NotBlank
