--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/project/entity/constant/ProjectSortBy.kt
@@ -1,6 +1,0 @@
-package team.themoment.datagsm.common.domain.project.entity.constant
-
-enum class ProjectSortBy {
-    ID,
-    NAME,
-}
