--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/project/entity/constant/ProjectStatus.kt
@@ -1,8 +1,0 @@
-package team.themoment.datagsm.common.domain.project.entity.constant
-
-enum class ProjectStatus(
-    val value: String,
-) {
-    ACTIVE("운영 중"),
-    ENDED("종료"),
-}
