+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/project/repository/custom/ProjectJpaCustomRepository.kt
@@ -3,9 +3,9 @@ package team.themoment.datagsm.common.domain.project.repository.custom
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import team.themoment.datagsm.common.domain.project.entity.ProjectJpaEntity
-import team.themoment.datagsm.common.domain.project.entity.constant.ProjectSortBy
-import team.themoment.datagsm.common.domain.project.entity.constant.ProjectStatus
 import team.themoment.datagsm.common.global.constant.SortDirection
+import team.themoment.datagsm.shared.domain.project.constant.ProjectSortBy
+import team.themoment.datagsm.shared.domain.project.constant.ProjectStatus
 
 interface ProjectJpaCustomRepository {
     fun searchProjectWithPaging(
