+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/student/dto/internal/StudentBulkUpdateDto.kt
@@ -1,9 +1,9 @@
 package team.themoment.datagsm.common.domain.student.dto.internal
 
 import team.themoment.datagsm.common.domain.club.entity.ClubJpaEntity
-import team.themoment.datagsm.common.domain.student.entity.constant.Major
-import team.themoment.datagsm.common.domain.student.entity.constant.Sex
-import team.themoment.datagsm.common.domain.student.entity.constant.StudentRole
+import team.themoment.datagsm.shared.domain.student.constant.Major
+import team.themoment.datagsm.shared.domain.student.constant.Sex
+import team.themoment.datagsm.shared.domain.student.constant.StudentRole
 
 data class StudentBulkUpdateDto(
     val id: Long,
