+++ datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/student/dto/request/UpdateStudentReqDto.kt
@@ -7,8 +7,8 @@ import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
-import team.themoment.datagsm.common.domain.student.entity.constant.Sex
-import team.themoment.datagsm.common.domain.student.entity.constant.StudentRole
+import team.themoment.datagsm.shared.domain.student.constant.Sex
+import team.themoment.datagsm.shared.domain.student.constant.StudentRole
 
 data class UpdateStudentReqDto(
     @field:Size(max = 30)
