--- datagsm-common/src/main/kotlin/team/themoment/datagsm/common/domain/student/entity/constant/StudentSortBy.kt
@@ -1,15 +1,0 @@
-package team.themoment.datagsm.common.domain.student.entity.constant
-
-enum class StudentSortBy {
-    ID,
-    NAME,
-    EMAIL,
-    STUDENT_NUMBER,
-    GRADE,
-    CLASS_NUM,
-    NUMBER,
-    MAJOR,
-    ROLE,
-    SEX,
-    DORMITORY_ROOM,
-}
